Move Ui class to ui folder
