+++ src/main/java/duke/command/AddCommand.java
@@ -10,7 +10,7 @@ import duke.task.Deadline;
 import duke.task.Event;
 import duke.task.Task;
 import duke.task.Todo;
-import duke.utils.Ui;
+import duke.ui.Ui;
 
 public class AddCommand extends Command{
