+++ src/main/java/duke/command/MarkCommand.java
@@ -3,7 +3,7 @@ package duke.command;
 import duke.constant.Message;
 import duke.database.DukeRepo;
 import duke.task.Task;
-import duke.utils.Ui;
+import duke.ui.Ui;
 
 public class MarkCommand extends Command {
