+++ src/main/java/duke/utils/DateUtil.java
@@ -32,6 +32,6 @@ public class DateUtil {
     }
 
     public static String dateToString(LocalDateTime date) {
-        return date.format(DateTimeFormatter.ofPattern("MMM d yyyy H:m"));
+        return date.format(DateTimeFormatter.ISO_LOCAL_DATE_TIME);
     }
 }
